BPE26-ref: extract runtime feature to dedicated file
